% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux.R
\docType{data}
\name{CC110315}
\alias{CC110315}
\title{Example of a nutrient data from the carbochemostat}
\format{\preformatted{'data.frame':	76 obs. of  8 variables:
 $ ID     : Factor w/ 76 levels "A1Ia1","A1IIa4",..: 29 30 31 32 1 4 7 10 13 15 ...
 $ NO2    : Named num  0.2 0.2 0.2 0.2 0.2 ...
  ..- attr(*, "names")= chr  NA NA NA NA ...
 $ NH4    : Named num  0.749 0.749 0.749 0.749 0.749 ...
  ..- attr(*, "names")= chr  NA NA NA NA ...
 $ PO4    : Named num  0.791 0.791 0.791 0.791 0.791 ...
  ..- attr(*, "names")= chr  NA NA NA NA ...
 $ NO3    : Named num  2.58 2.58 2.58 2.58 2.58 ...
  ..- attr(*, "names")= chr  NA NA NA NA ...
 $ TrueNO3: num  2.38 2.38 2.38 2.38 2.38 ...
 $ W__kg  : num  0.963 0.974 0.949 0.981 0.2 ...
 $ WR__kg : num  0.963 0.974 0.949 0.981 0.963 ...
}}
\usage{
CC110315
}
\description{
A data frame with NH3, NO2, NO3 and PO4 measurements fro a carbochemostat.
}
\keyword{data}

