% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ids.R
\name{checkId}
\alias{bottleId}
\alias{checkId}
\alias{cycleNbr}
\alias{itemNbr}
\alias{reactorId}
\alias{typeId}
\title{Convert from old to new chemostat Ids, and check these Ids}
\usage{
checkId(Ids)

reactorId(Ids)

itemNbr(Ids)

bottleId(Ids)

cycleNbr(Ids)

typeId(Ids)
}
\arguments{
\item{Ids}{The identifiers (either in old, or in new format).}
}
\description{
Ids are short code that fully identify a carbochemostat sample. It is in the
following format: Reactor # [A-Z], reactor or set [0-9],
bottle series # = [a-z], cycle [0-9][0-9], and addition or sample [a | s]. In
the old format, it was: Reactor or set = [R | A-C], reactor # = [1-4],
bottle series # = [I | II | III], addition or sample = [a | s], cycle = [0-9]+
}
\section{Functions}{
\itemize{
\item \code{reactorId}: Get Id of the reactor

\item \code{itemNbr}: Get the item number form its Id

\item \code{bottleId}: Get the bottle set Id

\item \code{cycleNbr}: Get the cycle number

\item \code{typeId}: Get the type (addition or sampling)
}}
\examples{
oldIds <- c("R1Is0", "R4Is0", "A1Ia1", "A2IIa4", "B3IIIs5",
            "C4IVs2", "B1Va2", "C4VIs6")
checkId(oldIds)
newIds <- c("A0a00s", "D0a00s", "A1a01a", "B1b04a", "C2c05s",
            "D3d02s", "A2e02a", "D3f06s")
checkId(newIds)
}

