% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flux.R
\name{conc}
\alias{conc}
\title{Concentration calculation for a variable from a chemdata object.}
\usage{
conc(var, data)
}
\arguments{
\item{var}{A character vector with the name of the variable(s) to use
for the calculations.}

\item{data}{A \code{chemdata} object, which default to \code{chemostat}.}
}
\value{
A \code{data.frame} with time and concentration for the variable(s)
}
\description{
Concentration of the solute (i.e., the variable) over time is extracted from
\code{data}
}
\details{
TODO: put some more info here...
}
\section{warning}{

Do provide duplicated measurements for the calculation of the error!
}
\examples{
\dontrun{
conc("NH3", data = chem01)
}
}
\seealso{
\code{\link{flux}}

Other calculation.functions: \code{\link{flux}}
}
\concept{
concentration
}

