% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.R
\name{flux}
\alias{flux}
\title{Flux calculation for a variable from a chemdata object.}
\usage{
flux(var, data)
}
\arguments{
\item{var}{A character vector with the name of the variable(s) to use
for the calculations.}

\item{data}{A \code{chemdata} object, which default to \code{chemostat}.}
}
\value{
A \code{data.frame} with time and flux for the variable(s)
}
\description{
With the carbochemostat, fluxes are calculated from the weight balance
from inputs and outputs, modulated by the concentration of the solute
(i.e., the variable).
}
\details{
TODO: put here the equations for the calculations of the flux...
}
\examples{
\dontrun{
flux("NH3", data = chem01)
}
}
\seealso{
\code{\link{conc}}, \url{http://econum.umons.ac.be}

Other calculation.functions: \code{\link{conc}}
}
\concept{
mass-balance flux
}

