% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pH.R
\name{pHDrift}
\alias{pHDrift}
\title{Compute the pH drift in seawater after partial equilibration with air}
\usage{
pHDrift(pHini, alk, elapsed, pCO2air = 400, tau = 560)
}
\arguments{
\item{pHini}{The initial pH in the seawater.}

\item{alk}{The initial alkalinity in mmol/kg.}

\item{elapsed}{Time elapsed waiting for equilibration in min.}

\item{pCO2air}{CO2 partial pressure of the gas phase in microatm}

\item{tau}{Cinetic constant for gas-liquid exchange of CO2 in the system.}
}
\value{
the final pH
}
\description{
Calculate the final pH of seawater after a given time considering
equilibration with a mixed gas phase considered as infinitely large
(constant pCO2air).
}
\examples{
pHDrift(pHini = 7.5, alk = 0.00230, elapsed = 60)
pHDrift(pHini = c(7, 8), alk = 0.00230, elapsed = c(60, 12000))
plot(1:500, pHDrift(pHini = 7, alk = 0.00230, elapsed = 1:500))
}

