\name{ENtest-package}
\alias{ENtest-package}
\alias{ENtest}
\docType{package}

\title{ EN test code }

\description{
  Various temporary test code for EcoNum.
}

\details{
  \tabular{ll}{
    Package: \tab ENtest\cr
    Type: \tab Package\cr
    Version: \tab 1.0-1\cr
    Date: \tab 2015-11-04\cr
    License: \tab MIT (see LICENSE file)\cr
    LazyLoad: \tab yes\cr
  }
}

\author{
  Philippe Grosjean

  Maintainer: Philippe Grosjean <phgrosjean@sciviews.org>
}

\references{
  EcoNum: http://econum.umons.ac.be/
}

\keyword{ package }

\concept{ Ecology }
