% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pH.R
\name{pHDrift}
\alias{pHDrift}
\title{Compute the pH of additions after partial equilibration with air}
\usage{
pHDrift(pHini, Alk, TimeElapsed__min, CO2ref = 400, tau = 560.2)
}
\arguments{
\item{pHini}{The initial pH.}

\item{Alk}{The initial alkalinity in mmol/kg-sw.}

\item{TimeElapsed__min}{Time elapsed waiting for equilibration}

\item{CO2ref}{CO2 partial pressure of the reference gas phase}

\item{tau}{Cinetic constant for gas-liquid exchange of CO2 in the systme}
}
\value{
the final pH
}
\description{
Adding a sample of seawater into another one, and calculate the final pH,
also considering equilibration with the gas phase during equilibration.
}
\examples{
pHDrift(pHini = 7.5, Alk = 0.00230, TimeElapsed = 60)
pHDrift(pHini = c(7, 8), Alk = 0.00230, TimeElapsed = c(60, 12000))
plot(1:500, pHDrift(pHini = 7, Alk = 0.00230, TimeElapsed = 1:500))
}

